Generic.Unification.Hinze

Description

Following the paper "Prolog control constructs in a functional setting" by Hinze, because I only want a clear denotational semantic for now. This is done as an experiment, and will not be part of the final package.

Documentation

class Monad m => Backtr m where

Important classes

Methods

flop :: m a

No solutions, corresponds to fail in prolog

amb :: m a -> m a -> m a

Non-deterministing choice between two branches

once :: m a -> m (Maybe a)

Encapsulation, find one solution

sols :: m a -> m [a]

Encapsulation, find all solutions

class Backtr m => Cut m where

Methods

cut :: m ()

like in prolog, cut discards choicepoints

call :: m a -> m a

to delimit the effect of a cut in a region

naf :: Backtr m => m a -> m ()

Negation as failure

naf2 :: Cut m => m a -> m ()

Subtly different negation as failure

only :: Cut m => a -> m a

Equivalent to once/1 in prolog. Note, the once function in the typeclass is the encapsulated version of this operation.

newtype BacktrT m a

Constructors

BacktrT

Fields

(>>-) :: forall ans. CPS a (m ans -> m ans)

newtype CutT m a

Constructors

CutT

Fields

(>>--) :: forall ans. CPS a (m (Answer m ans) -> m (Answer m ans))

newtype Logic a

Constructors

Logic

Fields

unLogic :: UnificationT (CutT Identity) a

runLogic :: Logic a -> [Either UnificationError (a, Substitution)]

evalLogic :: Logic a -> [a]

(===) :: Unifiable a => Term a -> Term a -> Logic (Term a)