Generic.Unification.Tutorial

Description

Use unification-sop if you want a mechanism to derive unification for your datatype in a generic way. When talking of unification, the reference implementation is the <http://hackage.haskell.org/package/unification-fd unification-fd> package. However that package has two major shortcomings in my opinion:

Your datatype has to be in the form of the fixpoint of a functor.

It's not immediately clear how to use it, and the documentation can be daunting.

that said, use that package if you are using unification for typechecking (which is the main use-case of that package) because it is surely more stable.

In this package we will see how to construct a different mechanism for unification which is strictly more expressive, letting us work with datatypes that are not in the fixpoint form.

Introduction

So, let's say you have a datatype, and you want to unify two terms:

data Foo

This is an example we'll use throughout the package

Constructors

FooI Int

FooS String Foo

The first step is deriving the generic instances. You can do that via

deriving (GHC.Generic, Generic, HasDatatypeInfo)

provided you have the DeriveGeneric and DeriveAnyClass extensions enabled (Generic and HasDatatypeInfo come from the generic-sop package).

And now, as your type is a tree which has only primitive types at the leaves, you get to construct generic terms.

A few examples of terms constructed in this way, with the prolog translation

ex1 :: Term Foo
ex1 = Var 1           -- X

ex2 :: Term Foo
ex2 = Con (FooI 3)          -- fooI 3

>>> ex1 :: Term Foo
Var 1

>>> ex4 :: Term Foo
fooS (Con "ciao") (Con (FooI 2))

ex1 :: Term Foo

ex2 :: Term Foo

Smart constructors

To make more easy working with terms, we propose the usage of smart constructors. For example, for our Foo datatype, we could define:

fooS :: Term String -> Term Foo -> Term Foo
fooS ts tf = Rec . SOP . S . Z $ ts :* tf :* Nil

fooI :: Term Int -> Term Foo
fooI ti = Rec . SOP . Z $ ti :* Nil

This constructs could be completely generated by the library (via generics or TH), but in this iteration of the library they have to be generated by the user. It is worth noting that the show instances I generated are geared towards this use-case, and that is automatic.

With there smart constructors defined, we can revisit previous examples:

ex3', ex4', ex5', ex5'var, ex5'var2 :: Term Foo
ex3' = fooI (Var 1)
ex4' = fooS (Con "ciao") (Con $ FooI 2)
ex5' = fooS (Con "ciao") (fooS (Var 1) (Con $ FooI 2))
ex5'var = fooS (Var 2) (fooS (Con "hey") (Con $ FooI 2))
ex5'var2 = fooS (Var 1) (fooS (Con "hey") (Con $ FooI 2))

Unification

Now let's have some examples of unification:

>>> unify (fooS (Var 2) (fooS (Con "hey") (Con $ FooI 2))) (fooS (Var 1) (fooS (Con "hey") (Con $ FooI 2)))
Right (fooS (Var 1) (fooS (Con "hey") (Con (FooI 2))))

Which is `fooS X (fooS "hey" (fooI 2)) == fooS Y (fooS "hey" (fooI 2))` in prolog-like notation. Take a better example here.

The result of unification can be Right and the substituted term, or Left and an explanation of the error.

Discussion

Here the differences with unification-fd

TODO Move the examples